chore: Refactor sorting algorithms and fix comparison counting

+++ src/BubbleSort.cpp
@@ -2,11 +2,15 @@
 
 #include <utility>
 
-void Sort::Bubble::Apply(int *data, int n) {
-	for (int step = 0; step < n - 1; step++) {
+void Sort::Bubble::Apply(int *data, int n)
+{
+	for (int step = 0; step < n - 1; step++)
+	{
 		bool swapped = false;
-		for (int i = 0; i < n - step - 1; i++) {
-			if (data[i] > data[i + 1]) {
+		for (int i = 0; i < n - step - 1; i++)
+		{
+			if (data[i] > data[i + 1])
+			{
 				std::swap(data[i], data[i + 1]);
 				swapped = true;
 			}
@@ -15,12 +19,16 @@ void Sort::Bubble::Apply(int *data, int n) {
 			break;
 	}
 }
-int Sort::Bubble::CountComparisons(int *data, int n) {
-	int count = 0;
-	for (int step = 0; ++count && step < n - 1; step++) {
+long long Sort::Bubble::CountComparisons(int *data, int n)
+{
+	long long count = 0;
+	for (int step = 0; ++count && step < n - 1; step++)
+	{
 		bool swapped = false;
-		for (int i = 0; ++count && i < n - step - 1; i++) {
-			if (++count && data[i] > data[i + 1]) {
+		for (int i = 0; ++count && i < n - step - 1; i++)
+		{
+			if (++count && data[i] > data[i + 1])
+			{
 				std::swap(data[i], data[i + 1]);
 				swapped = true;
 			}
