+++ src/BubbleSort.h
@@ -1,4 +1,5 @@
-namespace Sort::Bubble {
+namespace Sort::Bubble
+{
 	void Apply(int *data, int n);
-	int CountComparisons(int *data, int n);
+	long long CountComparisons(int *data, int n);
 }
