+++ src/MergeSort.h
@@ -1,5 +1,5 @@
 namespace Sort::Merge
 {
     void Apply(int *data, int start, int end);
-    int CountComparisons(int *data, int start, int end);
+    long long CountComparisons(int *data, int start, int end);
 }
