+++ src/SelectionSort.h
@@ -1,5 +1,5 @@
 namespace Sort::Selection
 {
     void Apply(int *data, int n);
-    int CountComparisons(int *data, int n);
+    long long CountComparisons(int *data, int n);
 }
